+++ dexterous_developer_cli/src/cross.rs
@@ -1,4 +1,6 @@
-use anyhow::bail;
+use anyhow::{bail, Context};
+
+use dexterous_developer_internal::Target;
 
 use crate::paths::{get_paths, CliPaths};
 
@@ -55,13 +57,24 @@ async fn setup_target(rust: &str) -> anyhow::Result<()> {
 
 pub const CROSS_TARGETS: &[&str] = &[
     #[cfg(target_os = "linux")]
-    "aarch64-unknown-linux-gnu",
+    Target::Linux.to_static(),
     #[cfg(target_os = "linux")]
-    "x86_64-unknown-linux-gnu",
+    Target::LinuxArm.to_static(),
     #[cfg(target_os = "linux")]
-    "x86_64-pc-windows-gnu",
-    #[cfg(target_os = "linux")]
-    "aarch64-apple-darwin",
-    #[cfg(target_os = "linux")]
-    "x86_64-apple-darwin",
+    Target::Windows.to_static(),
+    // #[cfg(target_os = "linux")]
+    // Target::Mac.to_static(),
+    // #[cfg(target_os = "linux")]
+    // Target::MacArm.to_static(),
 ];
+
+#[allow(clippy::single_match)]
+pub fn check_cross_requirements_installed(target: &Target) -> anyhow::Result<()> {
+    match target {
+        Target::Windows => {
+            which::which("x86_64-w64-mingw32-gcc").context("You need to install a Mingw-w64 cross compiler - some options are found here: https://www.mingw-w64.org/downloads/")?;
+        }
+        _ => {}
+    }
+    Ok(())
+}