+++ dexterous_developer_internal/src/hot/build_settings.rs
@@ -12,7 +12,7 @@ pub(crate) struct BuildSettings {
     pub features: String,
     pub target_folder: Option<PathBuf>,
     pub out_target: PathBuf,
-    pub build_target: Option<String>,
+    pub build_target: Option<crate::Target>,
     pub updated_file_channel: Option<tokio::sync::broadcast::Sender<HotReloadMessage>>,
 }
 
@@ -26,7 +26,7 @@ pub(crate) struct BuildSettings {
     pub features: String,
     pub target_folder: Option<PathBuf>,
     pub out_target: PathBuf,
-    pub build_target: Option<String>,
+    pub build_target: Option<crate::Target>,
 }
 
 impl ToString for BuildSettings {
@@ -100,6 +100,7 @@ impl TryFrom<&str> for BuildSettings {
         })
     }
 }
+
 
 #[cfg(feature = "cli")]
 use serde::{Deserialize, Serialize};