+++ dexterous_developer_internal/src/hot/mod.rs
@@ -185,3 +185,63 @@ pub async fn run_existing_library(library_path: std::path::PathBuf) -> anyhow::R
     run_from_file(library_paths, lib, null_watcher)?;
     Ok(())
 }
+
+#[cfg(feature = "cli")]
+pub fn compile_reloadable_libraries(
+    options: HotReloadOptions,
+    lib_dir: &std::path::Path,
+) -> anyhow::Result<std::path::PathBuf> {
+    use anyhow::Context;
+
+    let _ = env_logger::try_init();
+    let (mut settings, paths) = setup_build_settings(&options)?;
+    let lib_path = settings.lib_path.clone();
+
+    let lib_dir = if lib_dir.is_absolute() {
+        lib_dir.to_path_buf()
+    } else {
+        std::env::current_dir()?.join(lib_dir)
+    };
+    settings.out_target = lib_dir.clone();
+
+    if !lib_dir.exists() {
+        let _ = std::fs::create_dir_all(&lib_dir);
+    }
+
+    let lib_extension = lib_path
+        .extension()
+        .context("No extension for the library")?
+        .to_string_lossy();
+
+    for dir in paths.iter() {
+        if dir.as_path() != lib_dir && dir.exists() {
+            log::trace!("Checking lib path {dir:?}");
+            for file in (dir.read_dir()?).flatten() {
+                let path = file.path();
+                let extension = path
+                    .extension()
+                    .unwrap_or_default()
+                    .to_string_lossy()
+                    .to_string();
+                if path.is_file() && (extension == lib_extension) {
+                    let new_file = lib_dir.join(file.file_name());
+                    log::trace!("Moving {path:?} to {new_file:?}");
+                    std::fs::copy(path, new_file)?;
+                }
+            }
+        }
+    }
+
+    first_exec(&settings).context("Build failed")?;
+    let lib_path = lib_dir.join(lib_path.file_name().context("Lib must have a file name")?);
+    if !lib_path.exists() {
+        anyhow::bail!("{lib_path:?} wasn't built");
+    }
+
+    std::fs::copy(
+        &lib_path,
+        lib_path.with_extension(format!("{lib_extension}.backup")),
+    )?;
+
+    Ok(lib_dir.to_path_buf())
+}