+++ testing/dexterous_developer_tests/src/lib.rs
@@ -1,6 +1,9 @@
 mod utils;
 
-use std::{env, path::PathBuf};
+use std::{
+    env,
+    path::{Path, PathBuf},
+};
 
 use crate::utils::*;
 
@@ -239,7 +242,7 @@ async fn can_run_remote() {
 
     let mut process = client.run_client_cli("1234").await.unwrap();
 
-    process.is_ready();
+    process.is_ready().await;
 
     process.send("\n").expect("Failed to send empty line");
 
@@ -283,7 +286,7 @@ async fn can_update_assets() {
 
     let mut process = client.run_client_cli("2345").await.unwrap();
 
-    process.is_ready();
+    process.is_ready().await;
 
     process.send("\n").expect("Failed to send empty line");
 
@@ -300,6 +303,24 @@ async fn can_update_assets() {
 
     process.exit().await;
     host_process.exit().await;
+}
+
+async fn can_run_existing(path: &Path) {
+    let mut project = TestProject::existing_project(path, "can_run_existing").unwrap();
+
+    let mut process = project.run_existing().await.unwrap();
+
+    process.is_ready().await;
+
+    process.send("\n").expect("Failed to send empty line");
+
+    process.wait_for_lines(&["Ran Update"]).await;
+
+    process.send("\n").expect("Failed to send empty line");
+
+    process.wait_for_lines(&["Ran Update"]).await;
+
+    process.exit().await;
 }
 
 pub async fn run_tests() {
@@ -335,6 +356,19 @@ pub async fn run_tests() {
         "asset" => {
             println!("Can update asset");
             can_update_assets().await;
+        }
+        "existing" => {
+            println!("Can run existing assets");
+            let libs = args.next().expect("No next lib set");
+            let mut libs = PathBuf::from(libs);
+            if !libs.is_absolute() {
+                libs = std::env::current_dir().unwrap().join(libs);
+            }
+            if !libs.exists() || !libs.is_dir() {
+                panic!("libs should be a directory");
+            }
+            let libs = libs.canonicalize().unwrap();
+            can_run_existing(&libs).await;
         }
         _ => {
             eprintln!("{argument} is an invalid test");