+++ RELEASES.md
@@ -1,5 +1,13 @@
 # Release Notes
 
+## Version 0.0.9
+
+- Add tests for networked hot reload
+- Add `run`, `serve`, `remote`, `compile-libs`, `run-existing` and `install-cross` commands to the CLI
+- Add ability to compile reloadable libraries and load existing libraries
+- Add linux-to-windows cross compiled hot reload
+- Removed broken mold support
+
 ## Version 0.0.8
 
 - Add automated tests to validate hot reload across mac, windows & linux.